add clearDeposit func after send trx
