+++ src/pages/BorrowPage/components/BorrowForm/BorrowForm.module.scss
@@ -21,7 +21,7 @@
   text-transform: uppercase;
 
   @media screen and (max-width: 767px) {
-    margin: 30px 0 0;
+    margin: 28px 0 0;
   }
 }
 
@@ -55,7 +55,8 @@
   margin: 16px 0;
 
   @media screen and (max-width: 767px) {
-    margin: 4px 0;
+    margin-top: 8px;
+    margin-bottom: 4px;
   }
 }
